MARP-2054 Fix Sonar issues
