+++ marketplace-service/src/test/java/com/axonivy/market/service/impl/VersionServiceImplTest.java
@@ -8,14 +8,13 @@ import com.axonivy.market.entity.MavenArtifactVersion;
 import com.axonivy.market.entity.Metadata;
 import com.axonivy.market.entity.ProductJsonContent;
 import com.axonivy.market.enums.DevelopmentVersion;
-import com.axonivy.market.github.service.GHAxonIvyProductRepoService;
 import com.axonivy.market.model.MavenArtifactModel;
 import com.axonivy.market.model.VersionAndUrlModel;
 import com.axonivy.market.repository.MavenArtifactVersionRepository;
 import com.axonivy.market.repository.MetadataRepository;
 import com.axonivy.market.repository.ProductJsonContentRepository;
-import com.axonivy.market.repository.ProductRepository;
 import com.axonivy.market.util.MavenUtils;
+import com.axonivy.market.util.VersionUtils;
 import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Assertions;
@@ -46,13 +45,7 @@ class VersionServiceImplTest extends BaseSetup {
   private VersionServiceImpl versionService;
 
   @Mock
-  private GHAxonIvyProductRepoService gitHubService;
-
-  @Mock
   private MavenArtifactVersionRepository mavenArtifactVersionRepo;
-
-  @Mock
-  private ProductRepository productRepository;
 
   @Mock
   private ProductJsonContentRepository productJsonContentRepository;
@@ -168,12 +161,12 @@ class VersionServiceImplTest extends BaseSetup {
   @Test
   void testGetAllExistingVersions() {
     MavenArtifactVersion mockMavenArtifactVersion = new MavenArtifactVersion();
-    Assertions.assertTrue(CollectionUtils.isEmpty(MavenUtils.getAllExistingVersions(mockMavenArtifactVersion, false,
+    Assertions.assertTrue(CollectionUtils.isEmpty(VersionUtils.getAllExistingVersions(mockMavenArtifactVersion, false,
         StringUtils.EMPTY)));
     mockMavenArtifactVersion = getMockMavenArtifactVersionWithData();
-    Assertions.assertTrue(ObjectUtils.isNotEmpty(MavenUtils.getAllExistingVersions(mockMavenArtifactVersion, true,
+    Assertions.assertTrue(ObjectUtils.isNotEmpty(VersionUtils.getAllExistingVersions(mockMavenArtifactVersion, true,
         StringUtils.EMPTY)));
-    Assertions.assertTrue(CollectionUtils.isEmpty(MavenUtils.getAllExistingVersions(mockMavenArtifactVersion, false,
+    Assertions.assertTrue(CollectionUtils.isEmpty(VersionUtils.getAllExistingVersions(mockMavenArtifactVersion, false,
         StringUtils.EMPTY)));
   }
